Create README.md
This project will determine your Body-Mass Index. You just have to put your height and weight and your BMI will automatically be calculated and also your body type will be displayed on the screen. All you have to do is just start the XAMPP and in the folder of htdocs, create a folder named BMI and then run this program.